Fix compiler warning.
git-svn-id: https://scantailor.svn.sourceforge.net/svnroot/scantailor/trunk@131 80b70061-164d-0410-b7b7-e274ec932871

+++ imageproc/GaussBlur.cpp
@@ -37,7 +37,7 @@ static void transfer_pixels(
 	unsigned char* dest, int const dest_step, int const count)
 {
 	for (int i = 0; i < count; ++i) {
-		int sum = floor(*src1 + *src2 + 0.5);
+		int sum = (int)floor(*src1 + *src2 + 0.5);
 		if (sum > 255) {
 			sum = 255;
 		} else if (sum < 0) {
